backend/pyproject.toml
[project]
name = "schronisko-backend"
version = "0.1.0"
description = "schronisko-backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiocache>=0.12.3",
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.3",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
